% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataResource.R
\name{dataResource}
\alias{dataResource}
\title{Convenience functions to explore the datasets}
\usage{
dataResource()
}
\value{data object to provide desctiption of datasets}
\description{
Convenience functions to explore the datasets
}
\examples{
 
dataResource()
}
