\name{restfulSEData-package}
\alias{restfulSEData-package}
\alias{restfulSEData}
\title{restfulSEData - Example metadata for the "restfulSE" R package}
\description{
 This package includes SummarizedExperiment or RangedSummarizedExperiment
 instances from which assay data has been removed, so that it
 can be restored in real time from remote stores.
}



